Added user control for my code
