+++ BaseStation/MainWindow/MainWindow.xaml.cs
@@ -20,7 +20,7 @@ namespace HuskyRobotics.UI
         {
             InitializeComponent();
             DataContext = this;
-            DisplayMap();
+            
         }
 
         private void Button_Click(object sender, RoutedEventArgs e)
@@ -41,56 +41,7 @@ namespace HuskyRobotics.UI
             }
         }
 
-        private void DisplayMap()
-        {
-            // load in individual images
-            // var path = Path.Combine(System.Environment.CurrentDirectory, "Images", "map.jpg");
-            var uri = new System.Uri("F:/GitHub/2017-18/BaseStation/MainWindow/Images/map.jpg");
-            var bitmap = new BitmapImage(uri);
-            var hi = new BitmapImage(uri);
-            var image = new Image { Source = bitmap, Width = 300, Height = 300 };
-            var image2 = new Image { Source = bitmap, Width = 300, Height = 300 };
-            // put images to together 
-            Canvas.SetLeft(image, 0);
-            Canvas.SetTop(image, 0);
-            Canvas.SetLeft(image2, 300);
-            Canvas.SetTop(image2, 300);
-            canvas.Children.Add(image);
-            canvas.Children.Add(image2);
-            allImages[0] = image;
-            allImages[1] = image2;
-        }
-
-        private Point mousePosition;
-        private Image[] allImages = new Image[2];
-        private bool dragging = false;
-
-        private void CanvasMouseButtonDown(object sender, MouseButtonEventArgs e)
-        {
-            dragging = true;
-            mousePosition = e.GetPosition(canvas);
-        }
-
-        private void CanvasMouseButtonUp(object sender, MouseButtonEventArgs e)
-        {
-            dragging = false;
-        }
-
-        private void CanvasMouseMove(object sender, MouseEventArgs e)
-        {
-            if (dragging)
-            {
-                var position = e.GetPosition(canvas);
-                var offset = position - mousePosition;
-                mousePosition = position;
-                for (int i = 0; i < allImages.Length; i++)
-                {
-                    Canvas.SetLeft(allImages[i], Canvas.GetLeft(allImages[i]) + offset.X);
-                    Canvas.SetTop(allImages[i], Canvas.GetTop(allImages[i]) + offset.Y);
-                }
-
-            }
-        }
+        
     }
 }
