+++ BaseStation/MainWindow/MapDisplay.xaml
@@ -1,0 +1,17 @@
+﻿<UserControl x:Class="HuskyRobotics.UI.MapDisplay"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:HuskyRobotics.UI"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Canvas  ClipToBounds="True">
+        <Canvas 
+                    x:Name="canvas"
+                    MouseLeftButtonDown="CanvasMouseButtonDown"
+                    MouseLeftButtonUp="CanvasMouseButtonUp"
+                    MouseMove="CanvasMouseMove">
+        </Canvas>
+    </Canvas>
+</UserControl>
